move util.frozendict to types.frozendict
